Add files via upload
For this project, we will be analyzing data and statistics of NFL wide receivers. Football is the biggest sport in America and some of the biggest stars in the league are wide receivers. Receivers are the players who catch the passes the quarterback throws. They are the ones who score the majority of the touchdowns in the NFL. Due to the NFL’s huge success, millions of people now play fantasy football. This is a game where you draft NFL players, and you earn points based off their performance (reception = 1pt, 10 yards= 1pt, touchdown= 6 points). As you can see, touchdowns are worth a lot of points in fantasy football! So, we want players on our fantasy team who will score touchdowns. However, touchdowns are very volatile and vary year to year. So, how do we know what receivers will have the most touchdowns? Is there a way we can predict a receivers touchdowns based on his underlying stats and metrics? Using exploratory analysis and modeling, I will attempt to do just that in this project.